refactor(validation): refactoring validation schemas, refactoring handlers
