--- src/handlers/handleChange.js
@@ -1,6 +1,0 @@
-const handleChange =
-  (setFormData, formData) => (field) => (value) => {
-    setFormData({ ...formData, [field]: value });
-  };
-
-export default handleChange;
