--- src/handlers/handleOnChange.js
@@ -1,6 +1,0 @@
-const handleOnChange = (setEmpty, onChange) => (event) => {
-  setEmpty(true);
-  onChange(event.target.value);
-};
-
-export default handleOnChange;
