--- src/handlers/handleSubmit.js
@@ -1,5 +1,0 @@
-const handleSubmit = (formData) => {
-  console.log('Form data:', formData);
-};
-
-export default handleSubmit;
