+++ src/pages/RegisterPage/RegisterPage.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import AuthForm from '../../components/UI/AuthForm/AuthForm';
-import handleSubmit from '../../handlers/handleSubmit';
+import { handleSubmit } from '../../handlers/handlers';
 
 const RegisterPage = () => {
   const [formData, setFormData] = useState({
